requirements.txt
pandas>=1.3.0
seatable-api>=2.8.0
python-dotenv>=0.19.0
numpy>=1.21.0
openpyxl>=3.0.9